backend/requirements.txt
fastapi
uvicorn
torch
transformers
pandas
scikit-learn
tqdm
python-whois
dnspython
requests
tldextract
supabase
